Working verification test. Some last minute changes to component to re-abstract the store and various hacks just to get to this point. But working verification test!
