+++ src/Periturf.Tests/IdSvr4/ConfigurationComponentTests.cs
@@ -18,6 +18,7 @@ using IdentityServer4.Stores;
 using NUnit.Framework;
 using Periturf.Components;
 using Periturf.IdSvr4;
+using Periturf.IdSvr4.Configuration;
 using System;
 using System.Threading.Tasks;
 
@@ -34,16 +35,16 @@ namespace Periturf.Tests.IdSvr4
             const string clientId1 = "ClientID1";
             const string clientId2 = "ClientID2";
 
-            var component = new IdSvr4Component();
-            var clientStore = (IClientStore)component;
+            var store = new Store();
+            var clientStore = (IClientStore)store;
             IComponentConfigurator configurator1 = null;
             IComponentConfigurator configurator2 = null;
 
             var builder = A.Fake<IConfiugrationBuilder>();
             A.CallTo(() => builder.AddComponentConfigurator(componentName, A<Func<IdSvr4Component, IComponentConfigurator>>._))
-                .Invokes((string name, Func<IdSvr4Component, IComponentConfigurator> config) =>
+                .Invokes((string name, Func<Store, IComponentConfigurator> config) =>
                 {
-                    var configurator = config(component);
+                    var configurator = config(store);
 
                     if (configurator1 == null)
                         configurator1 = configurator;
